test/requirements.txt
pytest==7.4.3
pytest-asyncio==0.21.1
python-kubernetes==28.1.0
boto3==1.29.3
pyyaml==6.0.1
requests==2.31.0
python-consul2==0.1.5
pytest-timeout==2.2.0
cryptography==41.0.5
python-hcl2==4.3.2
